Cache Data
To cut down on downloading the same files over and over again, jwpub and
mp4 files are stored in a local cache. These files are used instead of
fresh copies if the checksum matches the one provided by the jw.org api.

Symlinks to the cached data are used rather than whole copies of the
files in the primary storage location to reduce local storage use.